python-code/requirements.txt
mysql-connector-python==8.0.26
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==4.0.2
flask-swagger-ui==3.36.0